feat: add orderbook fill market order ask

+++ src/matching_engine/engine.rs
@@ -1,4 +1,6 @@
 use std::collections::HashMap;
+use rust_decimal::Decimal;
+
 use super::orderbook::{Order, OrderBook};
 
 #[derive(Debug, Eq, PartialEq, Hash, Clone)]
@@ -36,7 +38,7 @@ impl MatchingEngine {
   pub fn place_limit_order(
     &mut self,
     pair: TradingPair,
-    price:f64,
+    price:Decimal,
     order:Order
   ) -> Result<(), String> {
     match self.orderbooks.get_mut(&pair) {
